REF: BulkRetryStrategy and utilities
